+++ .storybook/components/PicassoBook/Book.js
@@ -3,6 +3,7 @@ import Page from './Page'
 
 class PicassoBook extends Base {
   type = 'Book'
+  sections = []
 
   constructor(options = {}) {
     super(options)
@@ -15,11 +16,26 @@ class PicassoBook extends Base {
     return this
   }
 
-  createPage = (title, info, section) => {
-    const page = new Page({ title, info, section })
-    this.collection.push(page)
+  addSections = sections => {
+    this.sections = sections
+  }
 
-    return page
+  section = name => {
+    if (!this.sections.includes(name)) {
+      throw new Error(`
+Section with the name '${name}' was not found.
+You need to add this section to the list of available sections
+in '.storybook/config.js'.
+      `)
+    }
+
+    return {
+      createPage: (title, info) => {
+        const page = new Page({ title, info, section: name })
+        this.collection.push(page)
+        return page
+      }
+    }
   }
 
   createComponentDocs = (component, name, description, additionalDocs = {}) => {
@@ -29,7 +45,22 @@ class PicassoBook extends Base {
   connectToPage = connector => page => connector(page)
 
   generate() {
-    this.collection.forEach(page => page.generate())
+    // {
+    //   'Tutorials': 0,
+    //   'Components': 1,
+    //   ...
+    // }
+    const sectionsOrder = {}
+    for (let index in this.sections) {
+      sectionsOrder[this.sections[index]] = Number(index)
+    }
+
+    this.collection
+      .sort(
+        (page1, page2) =>
+          sectionsOrder[page1.section] - sectionsOrder[page2.section]
+      )
+      .forEach(page => page.generate())
   }
 }