+++ packages/picasso/src/EnvironmentBanner/story/index.jsx
@@ -2,7 +2,7 @@ import { EnvironmentBanner } from '../EnvironmentBanner'
 
 import PicassoBook from '~/.storybook/components/PicassoBook'
 
-const page = PicassoBook.createPage(
+const page = PicassoBook.section('Components').createPage(
   'EnvironmentBanner',
   'Banner on page top to highlight current environment'
 )