+++ packages/picasso/src/Grid/story/index.jsx
@@ -3,13 +3,12 @@ import { Grid } from '../Grid'
 
 import PicassoBook from '~/.storybook/components/PicassoBook'
 
-const page = PicassoBook.createPage(
+const page = PicassoBook.section('Layout').createPage(
   'Grid',
   `The layout element which is providing grid container functionality. 
-  The grid is built on top of flexbox functionality and the layout can 
-  be adjusted using the flexbox technics. The grid is a wrapper for the 
-  GridItem components.`,
-  'Layout'
+    The grid is built on top of flexbox functionality and the layout can 
+    be adjusted using the flexbox technics. The grid is a wrapper for the 
+    GridItem components.`
 )
 
 page