+++ packages/picasso/src/Radio/story/index.jsx
@@ -3,11 +3,10 @@ import { Radio } from '../Radio'
 
 import PicassoBook from '~/.storybook/components/PicassoBook'
 
-const page = PicassoBook.createPage(
+const page = PicassoBook.section('Forms').createPage(
   'Radio',
   `Radio buttons are best used when users need to select a single option from a set of unfamiliar choices. 
-  Radio buttons surface all the options and allow the user to compare choices before making a selection.`,
-  'Forms'
+    Radio buttons surface all the options and allow the user to compare choices before making a selection.`
 )
 
 page